refactor: 💡 rename form data
